+++ packages/api-pty/src/handlers.ts
@@ -3,6 +3,7 @@ import { apiGetPty } from './api.get-pty';
 import { apiListPty } from './api.list-pty';
 import { apiRemovePty } from './api.remove-pty';
 import { apiUpdatePty } from './api.update-pty';
+import { apiUploadPtyImage } from './api.upload-image';
 import { basePtyOs } from './orpc';
 
 const ptyHandlers = {
@@ -11,6 +12,7 @@ const ptyHandlers = {
   get: apiGetPty,
   update: apiUpdatePty,
   remove: apiRemovePty,
+  uploadImage: apiUploadPtyImage,
 };
 
 export { basePtyOs, ptyHandlers };